// packages/database/src/entities/Transaction.d.ts
import { BaseEntity } from './BaseEntity';
import { FuelType, PaymentMethod, PaymentStatus, TransactionStatus } from '@omc-erp/shared-types';
import { Station } from './Station';
import { Tank } from './Tank';
import { Pump } from './Pump';
import { User } from './User';
import { Customer } from './Customer';
import { Shift } from './Shift';
export declare class Transaction extends BaseEntity {
    tenantId: string;
    stationId: string;
    pumpId: string;
    tankId: string;
    attendantId: string;
    customerId: string;
    shiftId: string;
    fuelType: FuelType;
    quantityLiters: number;
    unitPrice: number;
    grossAmount: number;
    taxRate: number;
    taxAmount: number;
    serviceCharge: number;
    totalAmount: number;
    paymentMethod: PaymentMethod;
    paymentReference: string;
    paymentStatus: PaymentStatus;
    paymentProcessedAt: Date;
    receiptNumber: string;
    posReference: string;
    transactionTime: Date;
    status: TransactionStatus;
    temperature: number;
    density: number;
    loyaltyPointsAwarded: number;
    discountAmount: number;
    station: Station;
    tank: Tank;
    pump: Pump;
    attendant: User;
    customer: Customer;
    shift: Shift;
    calculateAmounts(): void;
    generateReceiptNumber(): void;
    isCompleted(): boolean;
    isPaid(): boolean;
    canBeRefunded(): boolean;
    getNetAmount(): number;
}
//# sourceMappingURL=Transaction.d.ts.map